ref: move didprov proxy to connect, remove from common
