+++ packages/common/src/index.ts
@@ -2,4 +2,3 @@ export { ThreeIDError, assert } from './errors.js'
 export * from './types.js'
 export * from './utils.js'
 export * from './network.js'
-export * from './didProviderWindowProxy.js'
