+++ packages/connect/package.json
@@ -47,6 +47,8 @@
     "@3id/connect-display": "^0.4.0-alpha.0",
     "@3id/window-auth-provider": "^0.4.0-alpha.0",
     "@ceramicnetwork/blockchain-utils-linking": "^2.0.0-alpha.0",
+    "@ceramicnetwork/rpc-window": "^0.6.0",
+    "dids": "^3.0.0-alpha.0",
     "rpc-utils": "^0.6.1",
     "rxjs": "^7.5.2"
   },
@@ -61,7 +63,6 @@
     "@types/store": "^2.0.2",
     "@types/url-parse": "^1.4.6",
     "caip": "^1.0.0",
-    "dids": "^3.0.0-alpha.0",
     "ipfs-repo": "^13.0.7",
     "uint8arrays": "^3.0.0"
   },
